Set stimIndices in quest structure for first trial
This was being set to 0, rather than the actual stimulus index.

It doesn’t look like this was actually used anywhere in the operation of quest, but it could be looked at by a user to see what the stimulus trial order was, and this would break for the first trial in that case.  Indeed, doing that is how I found this bug.