Add code to allow marginalizing out nuisance parameters from psychometric function
Sometimes we may not care whether we determine the values of some nuisance parameter in the model.  For example, we may not care about the lapse parameter value.  The fancy way to handle this is to minimize entropy of a parameter where the nuisance parameters have been summed out.

I added code to handle this, but have not yet tested it.  The new code does not get invoked unless you initialize with the ‘marginalize’ key value pair, so this should not break anything old.